+++ src/core/main.cpp
@@ -9,7 +9,7 @@ int main(int argc, char *argv[])
 
     Application app;
 
-    if (!app.Init(1280, 720, "Singularity Engine v0.10.4-alpha"))
+    if (!app.Init(1280, 720, "Singularity Engine v0.11.0-alpha"))
         return 1;
 
     app.Run();